refactor: rename canUnassignCurrentEvent to canUnassignAssigneesPopoverEvent for clarity and update bulk action types in related components
